backend-server/requirements.txt
wheel
setuptools
autopep8
Flask>=2.0
flask-cors>=3.0
pymongo[srv]>=3.0
python-dotenv>=0.1
requests>=2.0
PyGithub>=1.5
# tree-sitter